переместил некоторые функции в utils
